requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.5.2
langgraph==0.2.55
langchain>=0.3.17
langchain-core>=0.3.33
langchain-google-genai==2.0.8
chromadb==0.5.23
sentence-transformers==2.2.2
python-multipart==0.0.6
PyMuPDF==1.23.8
python-dotenv==1.0.0
redis>=5.0.0